[FEAT] update language based on URL lang query
## What is the feature?
When the location url has lang as search param, it triggers the english UI.
Otherwise, use the french UI.

## Why does the user need it?
When I share my portfolio to an english speaker user, he will get instantly the english UI.
Otherwise, he has to look at the UI, see the language button, click on it, then get what he wants: english content.
It gives a better user experience if this is automatic.

## How did you implement it?
When the app component loads, it looks at the current location url. If the lang query is inside, then, the hook state language is set to 'EN' otherwise, 'FR'.
Which switch the text content to english, great, but the language button style did not followed.
So, I inform the LanguageSelector component through the props if the lang search param exist.
If yes: after the first render, add the class 'active' to the language button.

Now, when loading the website with an url with lang as search param, it renders the english UI.