[FEAT] tap on project triggers hover style
## What is the feature?
When taping once on the project tile on mobile, it triggers the hover style made for desktop.

## Why is it needed?
The user might be confused if the user interface behave differently than the desktop one.
This feature gets the closest to the desktop behavior.

## How did you implement it?
I have added an onClick function on the project component wrapper that adds a class 'show' to the project component.
In the CSS, it triggers the hover style (only on devices that can't hover like mobiles) when the  show class exists.
Otherwise, the devices that can hover, works around the hover pseudo element to get the style.

Now, when taping on a project tile on mobile, it triggers the same style as when we hover the project tile on desktop.