[UPDATE] data in skills and the about me text
## What is the update?
Update the data in the skills section and the about me text content
## Why is it needed?
It was not up to date anymore

## How did you implement it?
I have added to the list of skills in data.js Agile, React Native, Linux (and more?)
I have also updated the story in the about me section to reflect more accurately my current professionnal experience.

Now, the skills section show the agile, Jira, React Native, Linux (and more?) items. And the story in about me is up to date.